+++ include/mqtt/broker/broker.hpp
@@ -308,7 +308,7 @@ public:
 
         ep.socket().lowest_layer().set_option(as::ip::tcp::no_delay(true));
         ep.set_auto_pub_response(false);
-        ep.set_async_notify(true);
+        ep.set_async_operation(true);
         ep.set_topic_alias_maximum(MQTT_NS::topic_alias_max);
         // Pass spep to keep lifetime.
         // It makes sure wp.lock() never return nullptr in the handlers below